[Q20-5] moved files into refactored package structure
and fixed menu and moved templates into correct folders
