+++ quadriga/src/main/java/edu/asu/diging/quadriga/core/data/CollectionRepository.java
@@ -1,9 +1,9 @@
-package edu.asu.diging.quadriga.core.mongo;
+package edu.asu.diging.quadriga.core.data;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import edu.asu.diging.quadriga.domain.elements.Collection;
+import edu.asu.diging.quadriga.core.model.Collection;
 
 @Repository
 public interface CollectionRepository extends MongoRepository<Collection, String>{
