+++ quadriga/src/main/java/edu/asu/diging/quadriga/core/service/CollectionManager.java
@@ -1,0 +1,8 @@
+package edu.asu.diging.quadriga.core.service;
+
+import edu.asu.diging.quadriga.core.model.Collection;
+
+public interface CollectionManager {
+
+    public Collection addCollection(String name, String description);
+}
