--- quadriga/src/main/java/edu/asu/diging/quadriga/service/ICollectionManager.java
@@ -1,8 +1,0 @@
-package edu.asu.diging.quadriga.service;
-
-import edu.asu.diging.quadriga.domain.elements.Collection;
-
-public interface ICollectionManager {
-
-    public Collection addCollection(Collection collection);
-}
