+++ quadriga/src/main/java/edu/asu/diging/quadriga/web/AddCollectionController.java
@@ -10,35 +10,31 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-import edu.asu.diging.quadriga.core.mongo.CollectionRepository;
-import edu.asu.diging.quadriga.domain.elements.Collection;
-import edu.asu.diging.quadriga.service.ICollectionManager;
+import edu.asu.diging.quadriga.core.service.CollectionManager;
 import edu.asu.diging.quadriga.web.forms.CollectionForm;
 
 @Controller
 public class AddCollectionController {
 
 	@Autowired
-	private ICollectionManager collectionManager;
+	private CollectionManager collectionManager;
 
 
 
 	@RequestMapping(value = "/auth/collections/add", method = RequestMethod.GET)
 	public String get(Model model) {
 		model.addAttribute("collectionForm", new CollectionForm());
-		return "admin/user/addCollection";
+		return "auth/addCollection";
 	}
 
 	@RequestMapping(value = "/auth/collections/add", method = RequestMethod.POST)
 	public String add(@Valid CollectionForm collectionForm, BindingResult result,
 			RedirectAttributes redirectAttrs) {
 		if (result.hasErrors()) {
-			return "admin/user/addCollection";
+			return "auth/addCollection";
 		}
-		Collection collection=new Collection();
-		collection.setName(collectionForm.getName());
-		collection.setDescription(collection.getDescription());
-		collectionManager.addCollection(collection);
+		
+		collectionManager.addCollection(collectionForm.getName(), collectionForm.getDescription());
 
 		redirectAttrs.addFlashAttribute("alert_type", "success");
 		redirectAttrs.addFlashAttribute("alert_msg", "Collection has been added.");
