+++ quadriga/src/main/webapp/WEB-INF/views/auth/addCollection.html
