+++ quadriga/src/test/java/edu/asu/diging/quadriga/core/service/impl/CollectionManagerImplTest.java
@@ -8,17 +8,16 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
-import edu.asu.diging.quadriga.core.mongo.CollectionRepository;
-import edu.asu.diging.quadriga.domain.elements.Collection;
-import edu.asu.diging.quadriga.service.impl.CollectionManager;
+import edu.asu.diging.quadriga.core.data.CollectionRepository;
+import edu.asu.diging.quadriga.core.model.Collection;
 
-public class CollectionManagerTest {
+public class CollectionManagerImplTest {
 
 	@Mock
 	private CollectionRepository collectionRepo;
 
 	@InjectMocks
-	private CollectionManager managerToTest;
+	private CollectionManagerImpl managerToTest;
 
 	@Before
 	public void setUp() {
@@ -32,7 +31,7 @@ public class CollectionManagerTest {
 		collection1.setName("name");
 		collection1.setDescription("testCollection");
 		Mockito.when(collectionRepo.save(collection1)).thenReturn(collection1);
-		Collection collection=managerToTest.addCollection(collection1);
+		Collection collection=managerToTest.addCollection(collection1.getName(), collection1.getDescription());
 		Mockito.verify(collectionRepo).save(collection);
 		Assert.assertNotNull(collection);
